pyproject.toml
[tool.poetry]
name = "pokedex"
version = "0.1.0"
description = "AI powered pokedex"
authors = ["Adam"]
license = "Open Domain"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
Flask = "^3.0.3"
Flask-Cors = "^5.0.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
